Changed the file names to be more appropiate. Added capture name class to main game file
